fix: 修复房产管理建筑面积字段映射问题
- 将HouseVO中的buildingAreaNum和usableArea字段类型从Double改为BigDecimal以正确匹配数据库decimal类型
- 修复HouseMapper.xml中resultMap的列映射，将buildingAreaNum的column从"building_area"改为"buildingAreaNum"以匹配SQL查询中的别名
- 添加调试日志以跟踪数据映射过程

+++ property-management-backend/src/main/java/com/hyu/property/domain/vo/HouseVO.java
@@ -3,6 +3,7 @@ package com.hyu.property.domain.vo;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * 房产视图对象 HouseVO
@@ -36,10 +37,10 @@ public class HouseVO implements Serializable {
     private String houseType;
 
     /** 建筑面积 */
-    private Double buildingAreaNum;
+    private BigDecimal buildingAreaNum;
 
     /** 使用面积 */
-    private Double usableArea;
+    private BigDecimal usableArea;
 
     /** 房产状态：1-空置 2-已售 */
     private Integer houseStatus;
